requirements.txt
langchain
langchain-ollama
langchain-chroma
pandas
streamlit
openai 
python-dotenv
PyPDF2
